+++ src/Ae.Gengo.Client/Entities/QuotedJobs.cs
@@ -6,6 +6,6 @@ namespace Ae.Gengo.Client.Entities
     [DataContract]
     public sealed class QuotedJobs : JobDictionaryContainer<JobQuote>
     {
-        public float TotalCredits => Jobs.Values.Sum(x => x.Credits);
+        public decimal TotalCredits => Jobs.Values.Sum(x => x.Credits);
     }
 }
